rename llm approaches
