+++ llm4pddl/approaches/__init__.py
@@ -1,7 +1,7 @@
 """Approaches module."""
 
 from llm4pddl.approaches.base_approach import BaseApproach
-from llm4pddl.approaches.llm_approaches import LLMOpenLoopApproach
+from llm4pddl.approaches.llm_open_loop_approach import LLMOpenLoopApproach
 from llm4pddl.approaches.pure_planning_approach import PurePlanningApproach
 from llm4pddl.flags import FLAGS
